% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{deriv}
\alias{deriv}
\title{Data derivation lookup.}
\format{A data frame with 55 observations and 2 variables:
\describe{
\item{deriv_id}{Unique identifer for deriv.}
\item{deriv}{Text description.}
}}
\usage{
deriv
}
\description{
This file (Table 11) provides information on how the nutrient values were
determined. The file contains the derivation codes and their descriptions
deriv.
}
\examples{
deriv

if (require("dplyr")) {
nutrient \%>\%
  select(nutr_id, deriv_id) \%>\%
  left_join(deriv)
}
}
\keyword{datasets}

