% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{food}
\alias{food}
\title{Food descriptions.}
\format{A data frame with 8463 observations and 14 variables:}
\usage{
food
}
\description{
To get information about nutrients, join with \code{\link{nutrient}}.
}
\examples{
food
}
\keyword{datasets}

