% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{food_group}
\alias{food_group}
\title{Food groups lookup table.}
\format{A data frame with 25 observations and 2 variables:
\describe{
\item{grp_id}{group id}
\item{group}{group name}
}}
\usage{
food_group
}
\description{
Food groups lookup table.
}
\examples{
food_group

if (require("dplyr")) {
  food \%>\%
    left_join(food_group) \%>\%
    select(food, group)
}
}
\keyword{datasets}

