% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{footnote}
\alias{footnote}
\title{Footnotes.}
\format{A data frame with 541 observations and 5 variables:
\describe{
\item{food_id}{Food identifier. Joins with \code{\link{food}}}
\item{seq}{Sequence number}
\item{type}{Type of footnote: D = footnote adding information to the food
  description; M = footnote adding information to measure description;
  N = footnote providing additional information on a nutrient value.
  If the type = N, the Nutr_No will also be filled in.}
\item{nutr_id}{Nutrient identifer. Joins with \code{\link{nutrient}}}
\item{footnote}{Footnote text}
}}
\usage{
footnote
}
\description{
This file (Table 13) contains additional information about the food item,
household weight, and nutrient value.
}
\examples{
footnote
}
\keyword{datasets}

