% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{nutrient_def}
\alias{nutrient_def}
\title{Nutrient definitions.}
\format{A data frame with 150 observations and 6 variables:
\describe{
\item{nutr_id}{Unique ID for nutrient}
\item{unit}{Units of measure}
\item{nutr_abbr}{International Network of Food Data Systems (INFOODS) tagname}
\item{nutr}{Name of nutrient/food component}
\item{precision}{Number of decimal places}
\item{seq}{Used to sort nutrient records in the same order as various
 reports produced from SR.}
}}
\usage{
nutrient_def
}
\description{
Nutrient definitions.
}
\examples{
nutrient_def
}
\keyword{datasets}

