% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{nutrient_source}
\alias{nutrient_source}
\title{Nutrient sources}
\format{A data frame with 213653 observations and 3 variables:
\describe{
\item{food_id}{Food identifier. Joins with \code{\link{nutrient}}}
\item{nutr_id}{Nutrient identifer. Joins with \code{\link{nutrient}}}
\item{ref_id}{Reference identifer. Joins with \code{\link{reference}}}
}}
\usage{
nutrient_source
}
\description{
This file (Table 14) is used resolve the many-to-many link between
nutrients with references.
}
\examples{
nutrient_source

if (require("dplyr")) {
  # Connect nutrients to references
  nutrient_source \%>\%
    inner_join(nutrient_def \%>\% select(nutr_id, nutr)) \%>\%
    inner_join(food \%>\% select(food_id, food)) \%>\%
    inner_join(reference)
}
}
\keyword{datasets}

