% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{reference}
\alias{reference}
\title{References (sources).}
\format{A data frame with 570 observations and 9 variables:
\describe{
\item{ref_id}{Reference identifier.}
\item{authors}{Comma delimited list of authors}
\item{title}{Title of paper}
\item{year}{Year published}
\item{journal}{Journal}
\item{vol_city}{}
\item{issue_state}{}
\item{start_page}{}
\item{end_page}{}
}}
\usage{
reference
}
\description{
References (sources).
}
\examples{
reference
}
\keyword{datasets}

