% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{source_type}
\alias{source_type}
\title{Source type lookup.}
\format{A data frame with 10 observations and 2 variables:
\describe{
\item{source_type_id}{Unique identifier for source type.}
\item{source_type}{Text description.}
}}
\usage{
source_type
}
\description{
This file (Table 10) contains codes indicating the type of data (analytical,
calculated, assumed zero, and so on) in the Nutrient Data file. To improve
the usability of the database and to provide values for the FNDDS, NDL staff
imputed nutrient values for a number of proximate components, total dietary
fiber, total sugar, and vitamin and mineral values.
}
\examples{
source_type

if (require("dplyr")) {
nutrient \%>\%
  select(nutr_id, source_type_id) \%>\%
  left_join(source_type)
}
}
\keyword{datasets}

