% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{source_type}
\alias{source_type}
\title{Source type lookup.}
\format{A data frame with 10 observations and 2 variables:
\describe{
\item{source_type_id}{Unique identifier for source type.}
\item{source_type}{Text description.}
}}
\usage{
source_type
}
\description{
Source type lookup.
}
\examples{
source_type

if (require("dplyr")) {
nutrient \%>\%
  select(nutr_id, source_type_id) \%>\%
  left_join(source_type)
}
}
\keyword{datasets}

